cast float values to uint16_t before passing to htons() in netjack
git-svn-id: svn+ssh://jackaudio.org/trunk/jack@3118 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ drivers/netjack/netjack_packet.c
@@ -759,13 +759,13 @@ render_jack_ports_to_payload_16bit (JSList *playback_ports, JSList *playback_src
     
                 for (i = 0; i < net_period_up; i++)
                 {
-                    packet_bufX[i] = htons ((floatbuf[i] + 1.0) * 32767.0);
+                    packet_bufX[i] = htons (((uint16_t)((floatbuf[i] + 1.0) * 32767.0)));
                 }
                 src_node = jack_slist_next (src_node);
             }
             else
                 for (i = 0; i < net_period_up; i++)
-                    packet_bufX[i] = htons((buf[i] + 1.0) * 32767.0);
+                    packet_bufX[i] = htons(((uint16_t)((buf[i] + 1.0) * 32767.0)));
         }
         else if (strncmp(portname, JACK_DEFAULT_MIDI_TYPE, jack_port_type_size()) == 0)
         {
